big test update(76%)

+++ src/test/java/ua/foxminded/pskn/universitycms/controller/CreateUserControllerTest.java
@@ -1,5 +1,12 @@
 package ua.foxminded.pskn.universitycms.controller;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.mockito.ArgumentMatchers.anyString;
+import static org.mockito.Mockito.doNothing;
+import static org.mockito.Mockito.eq;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.verify;
+
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
@@ -10,13 +17,8 @@ import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 import ua.foxminded.pskn.universitycms.dto.RoleDTO;
 import ua.foxminded.pskn.universitycms.dto.UserDTO;
 import ua.foxminded.pskn.universitycms.model.user.Role;
-import ua.foxminded.pskn.universitycms.model.user.User;
 import ua.foxminded.pskn.universitycms.service.user.RoleService;
 import ua.foxminded.pskn.universitycms.service.user.UserService;
-
-import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.Mockito.*;
 
 @WebMvcTest(CreateUserController.class)
 @WithMockUser(authorities = "ROLE_ADMIN")
@@ -35,8 +37,8 @@ class CreateUserControllerTest {
     void shouldCreateUser_AdminAuthorization() {
         UserDTO userDTO = new UserDTO();
         userDTO.builder().roleDTO(RoleDTO.builder().roleName("testrole").role(Role.builder().roleId(2).name("testrole").build()).build());
-        int groupId = 1;
-        int roleId = 2;
+        Long groupId = 1L;
+        Long roleId = 2L;
         RedirectAttributes redirectAttributes = mock(RedirectAttributes.class);
 
         doNothing().when(userService).createUserWithRole(userDTO, groupId, roleId);
